+++ dimas/src/com/liveliness.rs
@@ -4,6 +4,7 @@
 //! A `LivelinessSubscriber` can optional subscribe on a delete message.
 
 // region:		--- modules
+use core::time::Duration;
 use dimas_core::{
 	enums::{OperationState, TaskSignal},
 	error::Result,
@@ -11,7 +12,6 @@ use dimas_core::{
 };
 #[cfg(doc)]
 use std::collections::HashMap;
-use core::time::Duration;
 use tokio::task::JoinHandle;
 use tracing::info;
 use tracing::{error, instrument, warn, Level};